refactor: Redis 설정을 global로 통합하고 chat RedisConfig 내용 이전 후 삭제
